chore: refactor error framework

+++ omg-idl-gen/files/test-vectors/typedef_array_dim_2/expected.rs
@@ -1,4 +1,3 @@
 
-#[allow(dead_code)]
-#[allow(non_camel_case_types)]
-pub type Foo = [[u8;2 as usize];3+3 as usize];
+#[allow(dead_code, non_camel_case_types)]
+pub type Foo = [[u8;2_usize];3+3_usize];
